+++ test/features/save-state-feature.js
@@ -14,6 +14,30 @@ class MisbehavingAdapter extends MemoryAdapter {
   }
 }
 
+class AsyncAdapter extends MemoryAdapter {
+  constructor(mandatoryProp) {
+    super();
+    this.mandatoryProp = mandatoryProp;
+  }
+  upsert(type, key, value, options) {
+    if (type === STORAGE_TYPE_STATE) return this.upsertState(type, key, value, options);
+    return super.upsert(type, key, value, options);
+  }
+  upsertState(type, key, value, options) {
+    if (this.mandatoryProp && !options?.[this.mandatoryProp]) throw new StorageError(`mandatory ${this.mandatoryProp} is mandatory`);
+    return new Promise((resolve, reject) => {
+      process.nextTick(async () => {
+        try {
+          await super.upsert(type, key, value, options);
+          resolve();
+        } catch (err) {
+          reject(err);
+        }
+      });
+    });
+  }
+}
+
 Feature('save state', () => {
   Scenario('source with service task that saves state and then a timer and a message event, when messaged auto-save is enabled', () => {
     /** @type {MemoryAdapter} */
@@ -401,7 +425,7 @@ Feature('save state', () => {
     });
   });
 
-  Scenario('using saveState service', () => {
+  Scenario('using saveState service function', () => {
     /** @type {import('../../types/interfaces.js').IStorageAdapter} */
     let adapter;
     /** @type {ReturnType<testHelpers.horizontallyScaled>} */
@@ -558,7 +582,7 @@ Feature('save state', () => {
     });
   });
 
-  Scenario('process fails on manual save state app', () => {
+  Scenario('autosave is disabled and process run fails', () => {
     /** @type {MemoryAdapter} */
     let adapter;
     /** @type {ReturnType<testHelpers.horizontallyScaled>} */
@@ -572,7 +596,7 @@ Feature('save state', () => {
     });
 
     let deploymentName;
-    Given('a source matching scenario is deployed', async () => {
+    Given('a source with a script task that throws', async () => {
       deploymentName = 'volatile-process';
 
       await testHelpers.createDeployment(
@@ -594,7 +618,7 @@ Feature('save state', () => {
     let calledApp;
     let token;
     let response;
-    When('process is started with a volatile service function', async () => {
+    When('process is started', async () => {
       calledApp = apps.balance();
       errored = testHelpers.waitForProcess(calledApp, deploymentName).error();
 
@@ -621,21 +645,21 @@ Feature('save state', () => {
     });
   });
 
-  Scenario('adapter misbehaves', () => {
+  Scenario('process throws and an adapter with builtin reference error when updating', () => {
     /** @type {import('../../types/interfaces.js').IStorageAdapter} */
     let adapter;
     /** @type {ReturnType<testHelpers.horizontallyScaled>} */
     let apps;
     before(() => {
       adapter = new MisbehavingAdapter();
-      apps = testHelpers.horizontallyScaled(2, { adapter, autosaveEngineState: false });
+      apps = testHelpers.horizontallyScaled(2, { adapter, autosaveEngineState: true });
     });
     after(() => {
       apps?.stop();
     });
 
     let deploymentName;
-    Given('a source matching scenario is deployed', async () => {
+    Given('a source with a bad script', async () => {
       deploymentName = 'volatile-process';
 
       await testHelpers.createDeployment(
@@ -645,6 +669,12 @@ Feature('save state', () => {
         <definitions xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL"
           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
           <process id="bp" isExecutable="true">
+            <startEvent id="start">
+              <timerEventDefinition>
+                <timeDuration xsi:type="tFormalExpression">PT0.001S</timeDuration>
+              </timerEventDefinition>
+            </startEvent>
+            <sequenceFlow id="to-task" sourceRef="start" targetRef="task" />
             <scriptTask id="task" scriptFormat="javascript">
               <script>next(new Error('Expected'));</script>
             </scriptTask>
@@ -657,7 +687,7 @@ Feature('save state', () => {
     let calledApp;
     let token;
     let response;
-    When('process is started with a volatile service function', async () => {
+    When('process is started on app with autosave enabled', async () => {
       calledApp = apps.balance();
       errored = testHelpers.waitForProcess(calledApp, deploymentName).error();
 
@@ -666,8 +696,69 @@ Feature('save state', () => {
       token = response.body.id;
     });
 
-    Then('run fails', () => {
-      return errored;
+    Then('run fails with bad script error', async () => {
+      const err = await errored;
+      expect(err.message).to.equal('Expected');
+    });
+
+    And('engine is not running', () => {
+      expect(apps.getRunningByToken(token)).to.have.length(0);
+    });
+
+    When('attempting to get state of the failed run', async () => {
+      response = await apps.request().get('/rest/state/' + token);
+    });
+
+    Then('the failed run state is returned', () => {
+      expect(response.statusCode, response.text).to.equal(200);
+    });
+
+    When('resuming run', async () => {
+      response = await apps
+        .request()
+        .get('/rest/state/' + token)
+        .send({
+          id: 'Message_0',
+        });
+    });
+  });
+
+  Scenario('adapter throws on manual save state', () => {
+    /** @type {AsyncAdapter} */
+    let adapter;
+    /** @type {ReturnType<testHelpers.horizontallyScaled>} */
+    let apps;
+    before(() => {
+      adapter = new AsyncAdapter('prop');
+      apps = testHelpers.horizontallyScaled(2, { adapter, autosaveEngineState: false });
+    });
+    after(() => {
+      apps?.stop();
+    });
+
+    let deploymentName;
+    Given('a source matching scenario is deployed', async () => {
+      deploymentName = 'async-save-state-process';
+
+      await testHelpers.createDeployment(apps.balance(), deploymentName, saveStateResource);
+    });
+
+    let errored;
+    let app;
+    let token;
+    let response;
+    When('process is started with monitoring of middleware broker', async () => {
+      app = apps.balance();
+      errored = testHelpers.waitForProcess(app, deploymentName).error();
+
+      response = await request(app).post(`/rest/process-definition/${deploymentName}/start`).expect(201);
+
+      token = response.body.id;
+    });
+
+    Then('run fails with storage error', async () => {
+      const err = await errored;
+      expect(err.message).to.equal('mandatory prop is mandatory');
     });
 
     And('engine is not running', () => {
@@ -685,6 +776,87 @@ Feature('save state', () => {
 
     Then('not found is returned', () => {
       expect(response.statusCode, response.text).to.equal(404);
+    });
+
+    Given('a source that throws after manual save and a timer to facilitate tinkering with the adapter', async () => {
+      adapter.mandatoryProp = undefined;
+
+      await testHelpers.createDeployment(
+        apps.balance(),
+        deploymentName,
+        `<?xml version="1.0" encoding="UTF-8"?>
+        <definitions xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL"
+          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+          <process id="bp" isExecutable="true">
+            <serviceTask id="save" implementation="\${environment.services.saveState}" />
+            <sequenceFlow id="to-timer" sourceRef="save" targetRef="timer" />
+            <intermediateCatchEvent id="timer">
+              <timerEventDefinition>
+                <timeDuration xsi:type="tFormalExpression">PT10S</timeDuration>
+              </timerEventDefinition>
+            </intermediateCatchEvent>
+            <sequenceFlow id="to-task" sourceRef="timer" targetRef="task" />
+            <scriptTask id="task" scriptFormat="javascript">
+              <script>next(new Error('Expected'));</script>
+            </scriptTask>
+          </process>
+        </definitions>`
+      );
+    });
+
+    let timer;
+    When('process is started', async () => {
+      app = apps.balance();
+      timer = testHelpers.waitForProcess(app, deploymentName).timer();
+
+      response = await request(app).post(`/rest/process-definition/${deploymentName}/start`).expect(201);
+
+      token = response.body.id;
+    });
+
+    let timerMsg;
+    Then('timer is started', async () => {
+      timerMsg = await timer;
+    });
+
+    Given('adapter is changed so that it will throw', () => {
+      adapter.mandatoryProp = 'mandatory';
+    });
+
+    const messages = [];
+    const emittedErrors = [];
+    And('monitoring middleware broker for engine errors and app for emitted errors', () => {
+      app.locals.broker.subscribeTmp(
+        app.locals.middleware.middleware.name,
+        'engine.error',
+        (_, msg) => {
+          messages.push(msg);
+        },
+        { noAck: true }
+      );
+
+      app.on('bpmn/error', (err) => emittedErrors.push(err));
+    });
+
+    When('timer times out', () => {
+      errored = testHelpers.waitForProcess(app, token).error();
+
+      const [engine] = apps.getRunningByToken(token);
+      engine.environment.timers.executing.find((t) => t.owner.id === timerMsg.content.id).callback();
+    });
+
+    Then('run fails', () => {
+      return errored;
+    });
+
+    And('only error related to bpmn engine run error was published', () => {
+      expect(messages).to.have.length(1);
+      expect(messages[0].content.message).to.equal('Expected');
+    });
+
+    And('storage error was emitted on app', () => {
+      const [, err] = emittedErrors;
+      expect(err.message).to.equal('mandatory mandatory is mandatory');
     });
   });
 });